+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 # 2.5 - (something)
 ## 2.5.2
 - `play` command finally works (at least a bit)
+- Updated `queue` command
+- Added `playing` command
+- Added greetings
+- Commands in the webpanel look better (probably)
 
 ## 2.5.1
 - Fixed `permissions` command